Move components to proper place
